Share the same Lagrange basis among all levels.
